Make download feed tiles to queue incrementally
This should save memory for very large numbers of tiles, so we're not storing
them forever. Tries to insert in such a way as to give download threads a change
to insert failed tiles, but there MIGHT still be a dining philosophers type of
problem in rare cases (see TODO).